requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
alibabacloud-credentials==0.2.0
alibabacloud-dysmsapi20170525==2.0.3
alibabacloud-endpoint-util==0.0.3
alibabacloud-openapi-util==0.1.5
alibabacloud-tea==0.2.8
alibabacloud-tea-openapi==0.2.7
alibabacloud-tea-util==0.3.5
APScheduler==3.8.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopy==4.0.0
beautifulsoup4==4.10.0
blinker==1.4
Brotli==1.0.9
browsermob-proxy==0.8.0
captcha==0.3
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
charset-normalizer==2.0.6
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
daphne==3.0.2
Django==3.2.7
h11==0.12.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kaitaistruct==0.9
linecache2==1.0.0
lxml==4.6.3
msgpack==1.0.2
multidict==5.2.0
mysqlclient==2.0.3
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyExecJS==1.5.1
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyquery==1.4.3
PySocks==1.7.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
selenium==3.141.0
selenium-wire==4.5.3
service-identity==21.1.0
sip==6.1.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
traceback2==1.4.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
unittest2==1.1.0
urllib3==1.26.7
wsproto==1.0.0
yarl==1.7.0
zope.interface==5.4.0
zstandard==0.15.2
